refactor: move quote types to project types folder
