+++ app/types/types/quoteComponent.ts
